feat: add support to concurrency and fixed any things

+++ src/main/java/com/estebandev/minicloud/service/FileManagerService.java
@@ -1,5 +1,6 @@
 package com.estebandev.minicloud.service;
 
+import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.nio.file.FileAlreadyExistsException;
 import java.nio.file.Files;
@@ -7,6 +8,7 @@ import java.nio.file.Path;
 import java.nio.file.attribute.FileTime;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
+import java.util.concurrent.locks.ReentrantReadWriteLock;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -16,10 +18,9 @@ import org.springframework.core.io.Resource;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
-import com.estebandev.minicloud.component.MediatypeParser;
 import com.estebandev.minicloud.service.exception.FileIsNotDirectoryException;
-import com.estebandev.minicloud.service.exception.FileNotFoundException;
 import com.estebandev.minicloud.service.utils.FileData;
+import com.estebandev.minicloud.service.utils.FileManagerUtils;
 
 import lombok.RequiredArgsConstructor;
 
@@ -28,6 +29,7 @@ import lombok.RequiredArgsConstructor;
 public class FileManagerService {
     Logger logger = LoggerFactory.getLogger(this.getClass());
     private final UserService userService;
+    private final ReentrantReadWriteLock fileLock = new ReentrantReadWriteLock();
 
     @Value("${var.filepath}")
     private String pathString;
@@ -45,43 +47,54 @@ public class FileManagerService {
 
             Files.createDirectory(root);
         } catch (IOException e) {
-            logger.error("FATAL ERROR {}", e.getMessage(), e.getStackTrace());
+            logger.error("FATAL ERROR {}", e.getMessage());
             System.exit(1);
         }
     }
 
     public List<FileData> listFiles(String path)
             throws FileNotFoundException, FileIsNotDirectoryException, IOException {
-        verifyRootDirectory();
+        fileLock.readLock().lock();
+        try {
 
-        Path root = getRoot();
-        Path dir = root.resolve(path);
+            verifyRootDirectory();
 
-        if (!Files.exists(dir))
-            throw new FileNotFoundException();
+            Path root = getRoot();
+            Path dir = root.resolve(path);
 
-        if (!Files.isDirectory(dir))
-            throw new FileIsNotDirectoryException("The file is not directory");
+            if (!Files.exists(dir))
+                throw new FileNotFoundException();
 
-        return Files.list(dir)
-                .map(filePath -> {
-                    return FileData.builder()
-                            .fileName(filePath.getFileName().toString())
-                            .path(getRoot().relativize(filePath))
-                            .directory(Files.isDirectory(filePath))
-                            .build();
-                })
-                .toList();
+            if (!Files.isDirectory(dir))
+                throw new FileIsNotDirectoryException("The file is not directory");
+
+            return Files.list(dir)
+                    .map(filePath -> {
+                        return FileData.builder()
+                                .fileName(filePath.getFileName().toString())
+                                .path(getRoot().relativize(filePath))
+                                .directory(Files.isDirectory(filePath))
+                                .build();
+                    })
+                    .toList();
+        } finally {
+            fileLock.readLock().unlock();
+        }
     }
 
     private void makeDirectory(Path dirPath)
             throws FileAlreadyExistsException, IOException {
-        verifyRootDirectory();
+        fileLock.writeLock().lock();
+        try {
+            verifyRootDirectory();
 
-        if (Files.exists(dirPath))
-            throw new FileAlreadyExistsException("The already exist");
+            if (Files.exists(dirPath))
+                throw new FileAlreadyExistsException("The already exist");
 
-        Files.createDirectory(dirPath);
+            Files.createDirectory(dirPath);
+        } finally {
+            fileLock.writeLock().unlock();
+        }
     }
 
     public void makeDirectory(String dirPathString)
@@ -93,70 +106,85 @@ public class FileManagerService {
 
     public void makeDirectory(String dirPathString, String dirName)
             throws FileAlreadyExistsException, IOException {
-        makeDirectory(getRoot().resolve(dirPathString).resolve(formatName(dirName)));
+        makeDirectory(getRoot().resolve(dirPathString).resolve(FileManagerUtils.formatName(dirName)));
     }
 
     public void uploadFile(MultipartFile multipartFile, String dirPathString)
             throws IOException, FileIsNotDirectoryException, FileNotFoundException {
-        verifyRootDirectory();
-
-        Path dirPath = getRoot().resolve(dirPathString).normalize();
-
-        if (!Files.exists(dirPath))
-            throw new FileNotFoundException(dirPath);
-
-        if (!Files.isDirectory(dirPath))
-            throw new FileIsNotDirectoryException(dirPath + " is not directory");
-
-        String fileName = formatName(formatName(multipartFile.getOriginalFilename()));
-        Path filePath = dirPath.resolve(fileName);
-        if (Files.exists(filePath)) {
-            int dotPos = fileName.lastIndexOf('.');
-
-            StringBuilder fileNameBuilder = new StringBuilder();
-
-            double randomNumber = Math.floor(Math.random() * 10);
-            if (dotPos > -1) {
-                String name = fileName.substring(0, dotPos);
-                String format = fileName.substring(dotPos);
-                fileNameBuilder.append(name);
-                fileNameBuilder.append(randomNumber);
-                fileNameBuilder.append(format);
-            } else {
-                fileNameBuilder.append(fileName);
-                fileNameBuilder.append(randomNumber);
+        try {
+            if (!fileLock.writeLock().tryLock(5, TimeUnit.SECONDS)) {
+                throw new IOException("Upload is unable now");
             }
 
-            fileName = fileNameBuilder.toString();
-            filePath = dirPath.resolve(fileName);
-        }
+            try {
+                verifyRootDirectory();
 
-        multipartFile.transferTo(filePath);
+                Path dirPath = getRoot().resolve(dirPathString);
+
+                if (!Files.exists(dirPath))
+                    throw new FileNotFoundException("File does not exist");
+
+                if (!Files.isDirectory(dirPath))
+                    throw new FileIsNotDirectoryException(dirPath + " is not directory");
+
+                String fileName = FileManagerUtils.formatName(multipartFile.getOriginalFilename());
+                Path filePath = dirPath.resolve(fileName);
+                if (Files.exists(filePath)) {
+                    filePath = dirPath.resolve(FileManagerUtils.uniqueName(fileName));
+                }
+
+                multipartFile.transferTo(filePath);
+            } finally {
+                fileLock.writeLock().unlock();
+            }
+        } catch (InterruptedException e) {
+            Thread.currentThread().interrupt();
+            throw new IOException("Interrupt to try get lock", e);
+        }
     }
 
     public Path rename(String pathString, String newName) throws IOException {
+        try {
+            if (!fileLock.writeLock().tryLock(5, TimeUnit.SECONDS)) {
+                throw new IOException("Rename is unable now");
+            }
 
-        Path filePath = getRoot().resolve(pathString).normalize();
-        Path newFilePath = filePath.getParent().resolve(formatName(newName));
+            try {
 
-        Files.move(filePath, newFilePath);
-        return getRoot().relativize(newFilePath);
+                Path filePath = getRoot().resolve(pathString);
+                Path newFilePath = filePath.getParent().resolve(FileManagerUtils.formatName(newName));
+
+                Files.move(filePath, newFilePath);
+                return getRoot().relativize(newFilePath);
+            } finally {
+                fileLock.writeLock().unlock();
+            }
+        } catch (InterruptedException e) {
+            Thread.currentThread().interrupt();
+            throw new IOException("Interrupt to try get lock", e);
+        }
     }
 
     public Resource findFile(String pathString) throws FileNotFoundException, IOException {
-        verifyRootDirectory();
+        fileLock.readLock().lock();
+        try {
+            verifyRootDirectory();
 
-        Path filePath = getRoot().resolve(pathString).normalize();
+            Path filePath = getRoot().resolve(pathString);
 
-        if (!Files.exists(filePath))
-            throw new FileNotFoundException(filePath);
+            if (!Files.exists(filePath))
+                throw new FileNotFoundException("File not exists");
 
-        if (Files.isDirectory(filePath))
-            throw new IOException("The file is a directory");
+            if (Files.isDirectory(filePath))
+                throw new IOException("The file is a directory");
 
-        validateFile(filePath);
+            FileManagerUtils.validateFile(filePath);
 
-        return new FileSystemResource(filePath);
+            return new FileSystemResource(filePath);
+
+        } finally {
+            fileLock.readLock().unlock();
+        }
     }
 
     public FileData findFileData(String pathString) throws FileNotFoundException, IOException {
@@ -171,32 +199,44 @@ public class FileManagerService {
         return FileData.builder()
                 .fileName(fileName)
                 .path(path)
-                .mediaType(getMimeType(path))
-                .size(convertBytesToMegabytes(Files.size(path)))
+                .mediaType(FileManagerUtils.getMimeType(path))
+                .size(FileManagerUtils.convertBytesToMegabytes(Files.size(path)))
                 .directory(Files.isDirectory(path))
                 .editable(editable)
                 .build();
     }
 
     public void delete(String pathString) throws FileNotFoundException, IOException {
-        Path filePath = getRoot().resolve(pathString);
+        try {
+            if (!fileLock.writeLock().tryLock(5, TimeUnit.SECONDS)) {
+                throw new IOException("Renable is unable now");
+            }
+            try {
+                Path filePath = getRoot().resolve(pathString);
 
-        if (!Files.exists(filePath))
-            throw new FileNotFoundException();
-        if (!Files.isWritable(filePath))
-            throw new IOException("You do not have perms");
-        if (Files.isDirectory(filePath) && !listFiles(pathString).isEmpty())
-            throw new IOException("The directory is not empty");
+                if (!Files.exists(filePath))
+                    throw new FileNotFoundException();
+                if (!Files.isWritable(filePath))
+                    throw new IOException("You do not have perms");
+                if (Files.isDirectory(filePath) && !listFiles(pathString).isEmpty())
+                    throw new IOException("The directory is not empty");
 
-        Files.delete(filePath);
+                Files.delete(filePath);
+            } finally {
+                fileLock.writeLock().unlock();
+            }
+        } catch (InterruptedException e) {
+            Thread.currentThread().interrupt();
+            throw new IOException("Interrupt to try get lock", e);
+        }
     }
 
     public Path getRoot() {
         return Path.of(pathString);
     }
 
-    public long getLastModifiedDateInMinutes (String pathString) throws IOException {
-        Path filePath = getRoot().resolve(pathString).normalize();
+    public long getLastModifiedDateInMinutes(String pathString) throws IOException {
+        Path filePath = getRoot().resolve(pathString);
         FileTime fTime = Files.getLastModifiedTime(filePath);
         return fTime.to(TimeUnit.MINUTES);
     }
@@ -209,42 +249,8 @@ public class FileManagerService {
         return isExist(path) && Files.isDirectory(getRoot().resolve(path));
     }
 
-    public static void validateFile(Path filePath) throws IOException, FileNotFoundException {
-        if (!Files.exists(filePath)) {
-            throw new FileNotFoundException();
-        }
-        if (!Files.isRegularFile(filePath)) {
-            throw new IOException("Is not archive " + filePath);
-        }
-        if (!Files.isReadable(filePath)) {
-            throw new IOException("You do not have perms to read it: " + filePath);
-        }
-    }
-
-    public static Path getParent(String pathString) {
-        Path path = Path.of(pathString).getParent();
-        return path == null ? Path.of(".") : path;
-    }
-
-    public static String getMimeType(Path filePath) throws IOException {
-        String fileName = filePath.getFileName().toString();
-        int lastDotIndex = fileName.lastIndexOf('.');
-        if (lastDotIndex == -1 || lastDotIndex == fileName.length() - 1) {
-            return "directory"; 
-        }
-        return MediatypeParser.getMediaType(fileName.substring(lastDotIndex + 1).toLowerCase());
-    }
-
     public String getMimeType(String filePathString) throws IOException {
-        return getMimeType(getRoot().resolve(filePathString));
+        return FileManagerUtils.getMimeType(getRoot().resolve(filePathString));
     }
 
-    public static String formatName(String name) {
-        return name.trim().replaceAll("[ /%\\\\:*?\"'<>`]", "-");
-    }
-
-    public static double convertBytesToMegabytes(long bytes) {
-        double result = (double) bytes / (1024 * 1024);
-        return Math.round(result * 100.0) / 100.0;
-    }
 }
